+++ src/func/control/keyboard.h
@@ -15,18 +15,18 @@ typedef Func *(*keyboard_control_func)(int key, Func *input);
 
 const char *KEYBOARD_KEYS = "zsxdcvgbhnjm,l.;/"; // assumes US keyboard layout
 
-Func *keyboard_create(int semitones, keyboard_control_func control, Func *input)
+Func *keyboard_create(size_t semitones, keyboard_control_func control, Func *input)
 {
-    int count = (int)strlen(KEYBOARD_KEYS);
+    size_t count = strlen(KEYBOARD_KEYS);
     Func **array = (Func **)malloc_(count * sizeof(Func *));
     if (array == NULL)
     {
         return error_null(csErrorMemoryAlloc);
     }
-    for (int i = 0; i < count; i++)
+    for (size_t i = 0; i < count; i++)
     {
         char key = KEYBOARD_KEYS[i];
-        Func *pitched = pitch_(exp2(i / (double)semitones), input);
+        Func *pitched = pitch_(exp2((double)i / (double)semitones), input);
         array[i] = control(key, pitched);
     }
     Func *output = add_create(count, array);
