+++ src/func/effect/reverb.h
@@ -19,7 +19,7 @@ typedef struct
     size_t index;
 } ReverbContext;
 
-static double reverb_eval(__U int count, Gen **args, Eval *eval, void *context_)
+static double reverb_eval(__U size_t count, Gen **args, Eval *eval, void *context_)
 {
     ReverbContext *context = (ReverbContext *)context_;
     double tick = gen_eval(args[0], eval);
@@ -38,7 +38,7 @@ static double reverb_eval(__U int count, Gen **args, Eval *eval, void *context_)
     return output;
 }
 
-static void reverb_free(__U int count, void *context_)
+static void reverb_free(__U size_t count, void *context_)
 {
     ReverbContext *context = (ReverbContext *)context_;
     buffer_free(&context->buffer);
