+++ src/func/op/max.h
@@ -7,15 +7,14 @@
 #define CSYNTH_MAX_H
 
 #include <float.h>
-#include <stdarg.h>
 
 #include "../../core/func.h"
 #include "../../core/gen.h"
 
-static double max_eval(int count, Gen **args, Eval *eval, __U void *context)
+static double max_eval(size_t count, Gen **args, Eval *eval, __U void *context)
 {
     double max = -FLT_MAX;
-    for (int i = 0; i < count; i++)
+    for (size_t i = 0; i < count; i++)
     {
         double input = gen_eval(args[i], eval);
         if (max < input)
@@ -26,7 +25,7 @@ static double max_eval(int count, Gen **args, Eval *eval, __U void *context)
     return max;
 }
 
-Func *max_create(int count, Func **args)
+Func *max_create(size_t count, Func **args)
 {
     return func_create(NULL, max_eval, NULL, 0, NULL, FuncFlagNone, count, args);
 }
