+++ src/func/op/min.h
@@ -7,15 +7,14 @@
 #define CSYNTH_MIN_H
 
 #include <float.h>
-#include <stdarg.h>
 
 #include "../../core/func.h"
 #include "../../core/gen.h"
 
-static double min_eval(int count, Gen **args, Eval *eval, __U void *context)
+static double min_eval(size_t count, Gen **args, Eval *eval, __U void *context)
 {
     double min = FLT_MAX;
-    for (int i = 0; i < count; i++)
+    for (size_t i = 0; i < count; i++)
     {
         double input = gen_eval(args[i], eval);
         if (min > input)
@@ -26,7 +25,7 @@ static double min_eval(int count, Gen **args, Eval *eval, __U void *context)
     return min;
 }
 
-Func *min_create(int count, Func **args)
+Func *min_create(size_t count, Func **args)
 {
     return func_create(NULL, min_eval, NULL, 0, NULL, FuncFlagNone, count, args);
 }
