+++ src/func/op/neg.h
@@ -9,7 +9,7 @@
 #include "../../core/func.h"
 #include "../../core/gen.h"
 
-static double neg_eval(__U int count, Gen **args, Eval *eval, __U void *context)
+static double neg_eval(__U size_t count, Gen **args, Eval *eval, __U void *context)
 {
     double input = gen_eval(args[0], eval);
     return -input;
