+++ src/io/writer.h
@@ -12,9 +12,9 @@
 
 #define WRITER_BUFFER_SIZE 4096
 
-csError writer_write_channels_no_cleanup(double duration, FILE *file, int sample_rate, int channel_count, Func **channels)
+csError writer_write_channels_no_cleanup(double duration, FILE *file, size_t sample_rate, size_t channel_count, Func **channels)
 {
-    uint32_t sample_count = (uint32_t)(duration * sample_rate + 0.5);
+    uint32_t sample_count = (uint32_t)(duration * (double)sample_rate + 0.5);
     csError error = wav_header_write(sample_count, channel_count, file, sample_rate);
     if (error != csErrorNone)
     {
@@ -39,14 +39,14 @@ csError writer_write_channels_no_cleanup(double duration, FILE *file, int sample
     return csErrorNone;
 }
 
-csError writer_write_channels(double duration, FILE *file, int sample_rate, int count, Func **channels)
+csError writer_write_channels(double duration, FILE *file, size_t sample_rate, size_t count, Func **channels)
 {
     csError error = writer_write_channels_no_cleanup(duration, file, sample_rate, count, channels);
     cleanup_all();
     return error;
 }
 
-csError writer_write_file(double duration, const char *filename, int sample_rate, int channel_count, Func **channels)
+csError writer_write_file(double duration, const char *filename, size_t sample_rate, size_t channel_count, Func **channels)
 {
     FILE *file = fopen(filename, "wb");
     if (file == NULL)
@@ -66,9 +66,9 @@ csError writer_write_file(double duration, const char *filename, int sample_rate
     return csErrorNone;
 }
 
-csError write_channels(double duration, const char *filename, int count, Func **channels) { return writer_write_file(duration, filename, SAMPLE_RATE, count, channels); } /* writer_ */
-csError write(double duration, const char *filename, Func *input) { return write_channels(duration, filename, FUNCS(input)); }                                            /* writer_ */
-csError write_(double duration, Func *input) { return write(duration, DEFAULT_WAV_FILENAME, input); }                                                                     /* writer_ */
-csError write_stereo(double duration, const char *filename, Func *left, Func *right) { return write_channels(duration, filename, FUNCS(left, right)); }                   /* writer_ */
+csError write_channels(double duration, const char *filename, size_t count, Func **channels) { return writer_write_file(duration, filename, SAMPLE_RATE, count, channels); } /* writer_ */
+csError write(double duration, const char *filename, Func *input) { return write_channels(duration, filename, FUNCS(input)); }                                               /* writer_ */
+csError write_(double duration, Func *input) { return write(duration, DEFAULT_WAV_FILENAME, input); }                                                                        /* writer_ */
+csError write_stereo(double duration, const char *filename, Func *left, Func *right) { return write_channels(duration, filename, FUNCS(left, right)); }                      /* writer_ */
 
 #endif // CSYNTH_WRITER_H
